IFS-5424 ready for re-test.

+++ ifs-web-service/ifs-application-service/src/main/java/org/innovateuk/ifs/application/forms/sections/h2020costs/form/Horizon2020CostsForm.java
@@ -1,75 +1,75 @@
 package org.innovateuk.ifs.application.forms.sections.h2020costs.form;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public class Horizon2020CostsForm {
 
-    private BigDecimal labour;
-    private BigDecimal overhead;
-    private BigDecimal material;
-    private BigDecimal capital;
-    private BigDecimal subcontracting;
-    private BigDecimal travel;
-    private BigDecimal other;
+    private BigInteger labour;
+    private BigInteger overhead;
+    private BigInteger material;
+    private BigInteger capital;
+    private BigInteger subcontracting;
+    private BigInteger travel;
+    private BigInteger other;
 
-    public BigDecimal getLabour() {
+    public BigInteger getLabour() {
         return labour;
     }
 
-    public void setLabour(BigDecimal labour) {
+    public void setLabour(BigInteger labour) {
         this.labour = labour;
     }
 
-    public BigDecimal getOverhead() {
+    public BigInteger getOverhead() {
         return overhead;
     }
 
-    public void setOverhead(BigDecimal overhead) {
+    public void setOverhead(BigInteger overhead) {
         this.overhead = overhead;
     }
 
-    public BigDecimal getMaterial() {
+    public BigInteger getMaterial() {
         return material;
     }
 
-    public void setMaterial(BigDecimal material) {
+    public void setMaterial(BigInteger material) {
         this.material = material;
     }
 
-    public BigDecimal getCapital() {
+    public BigInteger getCapital() {
         return capital;
     }
 
-    public void setCapital(BigDecimal capital) {
+    public void setCapital(BigInteger capital) {
         this.capital = capital;
     }
 
-    public BigDecimal getSubcontracting() {
+    public BigInteger getSubcontracting() {
         return subcontracting;
     }
 
-    public void setSubcontracting(BigDecimal subcontracting) {
+    public void setSubcontracting(BigInteger subcontracting) {
         this.subcontracting = subcontracting;
     }
 
-    public BigDecimal getTravel() {
+    public BigInteger getTravel() {
         return travel;
     }
 
-    public void setTravel(BigDecimal travel) {
+    public void setTravel(BigInteger travel) {
         this.travel = travel;
     }
 
-    public BigDecimal getOther() {
+    public BigInteger getOther() {
         return other;
     }
 
-    public void setOther(BigDecimal other) {
+    public void setOther(BigInteger other) {
         this.other = other;
     }
 
     /* view logic. */
-    public BigDecimal getOrganisationFinanceTotal() {
+    public BigInteger getOrganisationFinanceTotal() {
         return labour
                 .add(overhead)
                 .add(material)
